Added base test class with common code
